UI-AA-52 update(badgeIcon): add types and cleanup
