+++ src/components/UiIconBadge/UiIconBadge.tsx
@@ -1,25 +1,9 @@
 import React from "react";
 import { ESize } from "../../_types/sizing";
 import cx from "classnames";
-import { type IconProp } from "@fortawesome/fontawesome-svg-core";
 import { UiIcon } from "../UiIcon/UiIcon";
-
-type BadgeSize = ESize.SM | ESize.MD | ESize.LG | ESize.XL;
-
-export enum EIconBadgeKind {
-	PRIMARY = "primary",
-	LIGHT = "light",
-	DARK = "dark",
-	WARNING = "warning",
-	ERROR = "error"
-}
-
-interface IUiIconBadgeProps {
-	name: IconProp;
-	size?: BadgeSize;
-	kind?: EIconBadgeKind;
-	children?: React.ReactNode;
-}
+import { EIconBadgeKind } from "./_types";
+import type { IUiIconBadgeProps } from "./_types";
 
 const badgeSizeClasses = {
 	[ESize.SM]: "p-xxs",
@@ -28,7 +12,6 @@ const badgeSizeClasses = {
 	[ESize.XL]: "p-lg"
 };
 
-// Map badge size to icon size
 const iconSizeMap = {
 	[ESize.SM]: ESize.SM,
 	[ESize.MD]: ESize.MD,
