+++ src/components/UiIconBadge/index.ts
@@ -1,1 +1,2 @@
 export { UiIconBadge } from "./UiIconBadge";
+export { EIconBadgeKind, type BadgeSize, type IUiIconBadgeProps } from "./_types";
